Removes Tiled functions form Main Menu Class
Removes the loading of Tiled maps to render the main menu. This method was not productive. Implements new graphics for the main menu using the default slick2d methodologies to produce a Main menu Image. Removes the polling for mouse location to highlight menu selections. Defines a function called isMouseHover that will return true when this is the case.