Changes UpdateTeams in Team class
Checks the x position of the car to see if its in the right or left half of the field. Dependent on which the algorithm will use the left or right ellipse to test if the next move will take the car outside of the boundaries. The middle rectangle area is always passed as this check must always occur.