Modifies UpdateCar function inside of cars class
adds a new Vector named next_move. The function calculates the next move then stores the newX and newY in this Vector. A call is then made to class CollidesHelper to a function CheckCollisions. If a collision type of "none" is returned the car is moved. Otherwise the Cars hit something and ProcessHit function is passed the collision type for processing .