Modifies method WallsCollide from using circles to using ellipse.
The values of the center_x and y and radius x and y for the ellipse are stored. These are  squared by muliplication and stored in another variable. The difference between the object x and center x and difference of object y and center y are calculated. These differences are then also squared and stored in x_center_diff and y_center_diff. The squared difference values from the center are then divided by their respective radius squared. The resulting ratios are summed and this value is later compared to 1. If greater than 1 then the point is outside the ellipse. First the point is evaluated against the center field rectangle bounds.