Adds switches when positioning cars on the field
Each case is decided upon how many players are on each team. if 1 then place at the screen_height/2. if 2 then place at intervals of 1/3. and if 3 then place them at 1/4 distance apart. The iterations take place in the y directions. The X is constant depending on which team the car corresponds.