Creates Debug variables in Main app class
is VelocityDebug shows velocity information of player controlled cars. Is DebugMovingDirection is a flag used during the creation of the quadrant class. Information on the facing directions and moving directions is calculated. isDebugField is a flag to turn on the wall collision ovals that layer on top of the field.