Creates the Quadrant Class
This class helps in determining the position of an object of the Cars class. Even though the car may be rotated at a certain angle they could be moving in another direction. Due to the confusion caused by the screen space coordinate system and the java trig functions this class was created to help with the busy work in calculations.