Fills out the Ball class definition
Balls of a radius, mass, a Circle object used when rendering the ball on the field. Creates a constructor that takes x and y as arguments. Constructor set the Circle object and adds a shape to the Entity of the same radius as the circle. Creates Render Function that takes Graphics as a argument. Sets the color of the graphics fill to orange and fills the ball then renders the JIG entity of the ball. A Function UpdateBall reports to console if the ball has been hit and a function Ball Hit is yet created but not implemented.