Creates a function which Generates the next move
Take Input as input and calls the ProcessInput function to see if any keys or buttons have been pressed. Variable next_move_direction is set based on the sin and cos of the turn angle in radians. Variable next_move_location is set to the current move plus the next move multiplied by the speed factor. PI is added to the Y component in order to satisify the flipped coordinate space.