Fills in Check world collisions function
This function takes each team and makes a list of them. It then iterates through the list and tests the extents of each car against a wall, ball, goalie, or goal. If a collision is made a report is filled out and placed in the cars list. ball has a list as well.