Alters UpdateBall method in Ball class
creates a speed limit for the ball so it can not fly out of control. 6.8f is the current limit. if speed is greater than this speed it is set to this value, and then reduced by 0.5f the next iteration. Create a line that updates the position where the sphere is being drawn on the screen. The update method then generates its next move, these will be used on teh next round of collision detection.