Fills in CheckWorld Collisions
CheckWorldCollisions takes arguments, Teams, Ball, Ellipse1, Ellipse2, Rectangle. These are our main colliders at the moment. All the cars are checked at each of their four extreme points for collisions. When a collision with wall is detected the car is basically replaced at a location far enough away from the wall. The collisions with the Ball are checked after wall. The direction vector of the ball and the car are added to determine the balls resulting trajectory. The balls speed is increased by the cars speed*2 since the ball is half the mass. A new next ball translation is created multiplying the ball speed by the bounced direction. The ball then regenerates its next moves using the Generate next move method.