Changes accelerate function to not use speed as a vector.
Speed is a magnitude value and accelerate would multiply neg one to the speed value in order to simulate reverse speed. This is now done through adding PI to the turn_rads whenever the accelerate or decelerate is first pressed. A flag reverse tracks the reversal and ensures the initial angle of direction remains as such.