Writes the Goalie class
Goalie extends Entity. There are two constants of MinY and MaxY which denote the range of a goalies motion in the Y direction. They can't move in the x direction.Goalies contain a rectangle that is drawn to screen representing their position in the field. Goalies have a method to track the ball. This method updates the goalies next position based on the ball position. Goalie also determines if the ball is stuck and then resets the ball to center.