Adds member variables to goalie class
Modifies speed to 1.5. creates two static variable GOLIE_MINY and GOLIE_MAXY that limit the goalie movement in the Y direction. Creates 3 vectors for movement, direction,translation, position. Creates a variable to hold the last known ball position aswell as initializes the last ball stuck timer to -1. Creates a variable for the is_stuck_timeout and sets it to 4.