Changes IsBallStuck method in Goalies class
finds the difference between the last known ball Y position and the current ball Ycoord. If that difference is zero and the timer is set to less than zero then the time needs to be noted as the first instance of not moving. The difference between the argument time and the last moved time is found until a difference greater than 5 is arrived. In this case the timer is reset and the function reports true. The function awaiting this true will then reposition the ball somewhere else.