Creates a rough way of tracking score through the play state
Creates two variables red_score and blue_score these are initialized to 0. Each time a goal is score teh IsGoal function returns a 1. this score is accumulated inside of each of the red and blue score functions. Next the scores are tested against the score limit. If either is over the limit they are tested against each other to declare the winner.